accountaFitBuddy/wsgi.py
"""
WSGI config for testingdjango project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/2.0/howto/deployment/wsgi/
"""

import os
from dotenv import load_dotenv
from django.core.wsgi import get_wsgi_application

BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))

os.environ.setdefault("DJANGO_SETTINGS_MODULE", "accountaFitBuddy.settings")
load_dotenv(os.path.join(BASE_DIR, '.env'))


application = get_wsgi_application()
