requirements.txt
Django
gunicorn
django-heroku
psycopg2
psycopg2-binary
whitenoise
dj-database-url
postgres
requests
Pillow
python-dotenv
